refactor(terminal): extract pure helpers into TerminalContainerView+Helpers
Moves static utility functions (compactPath, rebrandDisplayTitle, paneHeader*,
editorCommand, shellQuoted, matchesShortcut) and the EmbeddedTabSummary /
EmbeddedStatusSummary value types into a separate file. No behavior change —
all extracted members are pure or value types with no instance-state access.
