+++ Bellith/Views/TerminalContainerView.swift
@@ -98,26 +98,6 @@ final class TerminalContainerView: NSView, TerminalOverlayControllerHost, Termin
     /// to drive its own chrome (title bar, workspace rail, status bar).
     var onEmbeddedStateChanged: (() -> Void)?
 
-    /// Lightweight projection of the active tabs for the rebrand shell. Kept
-    /// flat / immutable so consumers can't mutate internal state.
-    struct EmbeddedTabSummary {
-        let id: UUID
-        let title: String
-        let paneCount: Int
-        let isSmart: Bool
-        let sourceIndex: Int
-    }
-
-    struct EmbeddedStatusSummary {
-        let muxName: String?
-        let paneCount: Int
-        let focusedPaneIndex: Int
-        let cwdDisplay: String?
-        let gitBranch: String?
-        let processDisplay: String?
-        let isBroadcasting: Bool
-    }
-
     /// Only terminal tabs land in the workspace rail — smart-panel tabs (file
     /// activity, process tree, etc.) get their own tools cluster eventually.
     var embeddedTabSummaries: [EmbeddedTabSummary] {
@@ -176,32 +156,6 @@ final class TerminalContainerView: NSView, TerminalOverlayControllerHost, Termin
         onEmbeddedStateChanged?()
     }
 
-    private static func rebrandDisplayTitle(for entry: TerminalTabEntry) -> String {
-        let trimmedTitle = entry.title.trimmingCharacters(in: .whitespacesAndNewlines)
-        if !trimmedTitle.isEmpty, !isGenericShellTitle(trimmedTitle) {
-            return trimmedTitle
-        }
-
-        let cwd = entry.focusedSurface?.currentCwd ?? entry.cwd
-        if let cwd, let name = rebrandWorkspaceName(from: cwd) {
-            return name
-        }
-        return trimmedTitle.isEmpty ? "session" : trimmedTitle
-    }
-
-    private static func rebrandWorkspaceName(from cwd: String) -> String? {
-        let trimmed = cwd.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return nil }
-        if trimmed == NSHomeDirectory() || trimmed == "~" { return "~" }
-        let normalized = trimmed.hasPrefix("~") ? NSString(string: trimmed).expandingTildeInPath : trimmed
-        let lastComponent = URL(fileURLWithPath: normalized).lastPathComponent
-        return lastComponent.isEmpty ? nil : lastComponent
-    }
-
-    private static func isGenericShellTitle(_ title: String) -> Bool {
-        let normalized = title.lowercased()
-        return ["zsh", "bash", "fish", "sh", "shell", "terminal"].contains(normalized)
-    }
     private let sidebarGlowLayer = CAGradientLayer()
     private let sidebarBridgeLayer = CAGradientLayer()
 
@@ -662,18 +616,6 @@ final class TerminalContainerView: NSView, TerminalOverlayControllerHost, Termin
             return true
         }
         return false
-    }
-
-    private static func matchesShortcut(
-        _ event: NSEvent,
-        key: String,
-        command: Bool = false,
-        shift: Bool = false,
-        option: Bool = false,
-        control: Bool = false
-    ) -> Bool {
-        let shortcut = KeyShortcut(key: key, command: command, shift: shift, option: option, control: control)
-        return shortcut.matches(event: event)
     }
 
     private func removeEventMonitors() {
@@ -1455,20 +1397,6 @@ final class TerminalContainerView: NSView, TerminalOverlayControllerHost, Termin
 
     private var shouldShowStatusBar: Bool {
         dependencies.settings.showStatusBar && statusBar.hasVisibleContent
-    }
-
-    private static func compactPath(_ path: String?) -> String? {
-        guard let path, !path.isEmpty else { return nil }
-        let home = NSHomeDirectory()
-        let normalized: String
-        if path.hasPrefix(home) {
-            normalized = "~" + path.dropFirst(home.count)
-        } else {
-            normalized = path
-        }
-        let pieces = normalized.split(separator: "/", omittingEmptySubsequences: false)
-        guard pieces.count > 3 else { return String(normalized) }
-        return pieces.suffix(3).joined(separator: "/")
     }
 
     private func refreshGitHubStatusAsync(cwd: String) {
@@ -2947,24 +2875,6 @@ final class TerminalContainerView: NSView, TerminalOverlayControllerHost, Termin
         }
     }
 
-    /// Compact pane title for the header — first word of the foreground process,
-    /// falling back to "zsh" so the header always has a label.
-    private static func paneHeaderTitle(from presentation: ForegroundProcessPresentation?) -> String {
-        guard let presentation else { return "zsh" }
-        let firstSegment = presentation.text.split(separator: " ").first.map(String.init) ?? presentation.text
-        return firstSegment.isEmpty ? "zsh" : firstSegment
-    }
-
-    /// A pane is "running" (warning-tinted dot) whenever a non-shell foreground
-    /// process is present — matches the design's distinction between idle shells
-    /// and active long-runners (dev servers, watchers, REPLs).
-    private static func paneHeaderIsRunning(from presentation: ForegroundProcessPresentation?) -> Bool {
-        guard let presentation else { return false }
-        let lower = presentation.text.lowercased()
-        let shellNames = ["zsh", "bash", "fish", "sh", "dash", "ksh"]
-        return !shellNames.contains(where: { lower == $0 || lower.hasPrefix("\($0) ") })
-    }
-
     private func makeSurface(tabId: UUID, app: TerminalApp, context: TerminalContext) -> TerminalSurfaceView {
         let surface = TerminalSurfaceView(app: app)
         surface.terminalContext = context
@@ -3077,22 +2987,6 @@ final class TerminalContainerView: NSView, TerminalOverlayControllerHost, Termin
         sessionCoordinator.restoreWorkingDirectory(cwd, on: surface)
     }
 
-    private static func editorCommand(for fileURL: URL) -> String {
-        let escapedPath = shellQuoted(fileURL.path)
-        return """
-        if [ -n "${EDITOR:-}" ]; then
-          eval "$EDITOR \(escapedPath)"
-        elif [ -n "${VISUAL:-}" ]; then
-          eval "$VISUAL \(escapedPath)"
-        else
-          open -t \(escapedPath)
-        fi
-        """
-    }
-
-    private static func shellQuoted(_ value: String) -> String {
-        "'\(value.replacingOccurrences(of: "'", with: "'\\''"))'"
-    }
 }
 
 // MARK: - Broadcast Badge
